+++ PROYECTO-UNIDAD2/src/producto/Categoria.java
@@ -1,0 +1,4 @@
+package producto;
+
+public class Categoria {
+}
